requirements.txt
langchain-cli==0.0.20
pydantic==1.10.13
nomic==2.0.16
pinecone-client==2.2.4
torch==2.1.2
sentence-transformers==2.2.2
transformers==4.36.2
datasets==2.15.0
scikit-learn==1.3.2
seaborn==0.13.0
umap-learn==0.5.5
hdbscan==0.8.33
bertopic==0.16.0
pandas==2.1.4
numpy==1.26.3
redis==5.0.1
jupyterlab==4.0.10
ipykernel==6.28.0
